refactor(templates): drop the v2 suffix from the proposal package
Second of the four per-family suffix drops. Move
templates.proposal.v2.presets to templates.proposal.presets and rewrite
every import and doc reference across the engine, tests, examples, and
benchmarks. Pure rename — the visual parity test passes against the
unchanged committed baseline, so the render is identical.

Tests: ./mvnw verify javadoc:javadoc -pl . — 1378 tests, 0 failures, javadoc
clean; examples + benchmarks compile; perf-smoke + examples-generation smoke
(85) green.
