+++ src/main/java/com/demcha/compose/document/templates/data/proposal/package-info.java
@@ -1,5 +1,5 @@
 /**
  * Shared, render-neutral proposal document specs and supporting data
- * records, consumed by the layered {@code proposal.v2} presets.
+ * records, consumed by the layered {@code proposal.presets} presets.
  */
 package com.demcha.compose.document.templates.data.proposal;
